Component moved to utils/components
- all references updated
