+++ src/chime_dash/app/components/additions.py
@@ -18,7 +18,7 @@ from penn_chime.parameters import Parameters
 from chime_dash.app.utils.templates import df_to_html_table
 from chime_dash.app.services.plotting import plot_dataframe
 
-from chime_dash.app.components.base import Component
+from chime_dash.app.utils.components import Component
 
 
 class Additions(Component):
