+++ src/chime_dash/app/components/definitions.py
@@ -8,7 +8,7 @@ from dash.development.base_component import ComponentMeta
 
 from dash_core_components import Markdown
 
-from chime_dash.app.components.base import Component
+from chime_dash.app.utils.components import Component
 
 
 class Definitions(Component):
