+++ src/chime_dash/app/components/intro.py
@@ -13,7 +13,7 @@ from dash_core_components import Markdown
 from penn_chime.parameters import Parameters
 
 from chime_dash.app.utils.templates import read_localization_markdown
-from chime_dash.app.components.base import Component
+from chime_dash.app.utils.components import Component
 
 LOCALIZATION_FILE_1 = "intro.md"
 LOCALIZATION_FILE_2 = "tool-details.md"
