+++ src/chime_dash/app/components/menu.py
@@ -8,7 +8,7 @@ from typing import List
 from dash.development.base_component import ComponentMeta
 import dash_bootstrap_components as dbc
 
-from chime_dash.app.components.base import Component
+from chime_dash.app.utils.components import Component
 
 
 class Menu(Component):
